[INTERNAL] Splitting test for seeds module.
This hopefully lowers compilation memory requirements for Travis builds.
